bdf:  - deck splitter now handles multiple spaces in label
bdf_vectorized3:
 - fixing BNDGRID writing bug (BNDGRID1 doesn't exist)
 - fixed GUI handling of nan concentrated mass
 - adding SEUSET, SEUSET1, MASSSET, CBEAM3
 - more int64 support
 - more allow_nan_thickness/mass support
 - better MONDSP1 parsing (tabs & comp field)

abaqus_to_nastran:
 - support for s3, s4, s3r, s4r
 - fixing element id mapping bug (didn't work when there was no overlapping)

+++ pyNastran/dev/bdf_vectorized3/cards/elements/fast.py
@@ -113,7 +113,7 @@ class CFAST(Element):
 
         fast_type = np.zeros(ncards, dtype='|U4')
         ids = np.zeros((ncards, 2), dtype=idtype)
-        fastener_node = np.zeros(ncards, dtype='int32')
+        fastener_node = np.zeros(ncards, dtype=idtype)
         fastener_xyz = np.zeros((ncards, 3), dtype='float64')
         for icard, card in enumerate(self.cards):
             (eid, pid, fast_typei, idsi, nodesi, gs, xyz, comment) = card
